requirements.txt
langchain_google_genai
langchain_core
langgraph
pydub
soundfile
librosa
ffmpeg
google-genai
